+++ patzilla/navigator/settings.py
@@ -2,45 +2,132 @@
 # (c) 2014-2017 Andreas Motl, Elmyra UG
 import json
 import logging
-from pyramid.settings import asbool         # Keep this for h.asbool()
+from pyramid.exceptions import ConfigurationError
 from pyramid.threadlocal import get_current_request, get_current_registry
 from patzilla.navigator.util import dict_prefix_key, dict_merge
+from patzilla.util.config import read_config, read_list, asbool
 from patzilla.util.date import datetime_isoformat, unixtime_to_datetime
 from patzilla.util.python import _exception_traceback
 from patzilla.util.data.container import SmartBunch
 
 log = logging.getLogger(__name__)
 
-class Bootstrapper(object):
-    """all parameter fiddling in one single place :-)"""
 
-    # TODO: refactor this out of helpers.py, just import here
-    # TODO: refactor IpsuiteNavigatorConfig.defaults here as well, trim down config.js
+class GlobalSettings(object):
+
+    def __init__(self):
+        self.registry = get_current_registry()
+        self.configfile = self.registry.settings['CONFIG_FILE']
+        self.application_settings = self.get_application_settings()
+        self.datasource_settings  = self.get_datasource_settings()
+        self.vendor_settings      = self.get_vendor_settings()
+
+    def get_application_settings(self):
+        # Read configuration file to get global settings
+        # TODO: Optimize: Only read once, not on each request!
+        return read_config(self.configfile, kind=SmartBunch)
+
+    def get_datasource_settings(self):
+
+        # Container for datasource settings
+        datasource_settings = SmartBunch({
+            'datasources': [],
+            'datasource': SmartBunch(),
+            'total': SmartBunch.bunchify({'fulltext_countries': [], 'details_countries': []}),
+            })
+
+        # Read datasource settings from configuration
+        datasource_settings.datasources = read_list(self.application_settings.get('ip_navigator', {}).get('datasources'))
+        datasource_settings.protected_fields = read_list(self.application_settings.get('ip_navigator', {}).get('datasources_protected_fields'))
+        for datasource in datasource_settings.datasources:
+            settings_key = 'datasource_{name}'.format(name=datasource)
+            datasource_info = self.application_settings.get(settings_key, {})
+            datasource_info['fulltext_enabled'] = asbool(datasource_info.get('fulltext_enabled', False))
+            datasource_info['fulltext_countries'] = read_list(datasource_info.get('fulltext_countries', ''))
+            datasource_info['details_enabled'] = asbool(datasource_info.get('details_enabled', False))
+            datasource_info['details_countries'] = read_list(datasource_info.get('details_countries', ''))
+            datasource_settings.datasource[datasource] = SmartBunch.bunchify(datasource_info)
+
+            # Aggregate data for all countries
+            datasource_settings.total.fulltext_countries += datasource_info['fulltext_countries']
+
+        return datasource_settings
+
+    def get_vendor_settings(self):
+
+        # Container for vendor settings
+        vendor_settings = SmartBunch({
+            'vendors': [],
+            'vendor': SmartBunch(),
+        })
+
+        # Read vendor settings from configuration
+        try:
+            vendor_settings.vendors = read_list(self.application_settings.ip_navigator.vendors)
+            assert vendor_settings.vendors
+        except:
+            raise ConfigurationError('No vendor configured in "{configfile}"'.format(configfile=self.configfile))
+
+        for vendor in vendor_settings.vendors:
+
+            settings_key = 'vendor_{name}'.format(name=vendor)
+            if settings_key not in self.application_settings:
+                raise ConfigurationError('Vendor "{vendor}" not configured in "{configfile}"'.format(
+                    vendor=vendor, configfile=self.configfile))
+
+            vendor_info = self.application_settings.get(settings_key, {})
+            for key, value in vendor_info.iteritems():
+                vendor_info[key] = value.decode('utf-8')
+
+            if 'hostname_matches' in vendor_info:
+                vendor_info.hostname_matches = read_list(vendor_info.hostname_matches)
+
+            vendor_settings.vendor[vendor] = SmartBunch.bunchify(vendor_info)
+
+        return vendor_settings
+
+
+class RuntimeSettings(object):
+    """
+    All runtime parameters in one single place
+    """
+
+    # TODO: Refactor IpsuiteNavigatorConfig.defaults here as well, trim down config.js
+
+    development_hosts = [
+        'localhost',
+        'offgrid',
+    ]
 
     staging_hosts = [
         'localhost',
         'offgrid',
         'patentsearch-develop.elmyra.de',
         'patentsearch-staging.elmyra.de',
-        ]
+    ]
 
     def __init__(self):
+
+        self.request = get_current_request()
+        self.registry = get_current_registry()
+
         self.config = self.config_parameters()
         self.theme = self.theme_parameters()
         self.beta_badge = '<div class="label label-success label-large do-not-print">BETA</div>'
 
+    def asdict(self):
+        return self.__dict__.copy()
+
     def environment(self):
         """create default environment"""
 
-        request = get_current_request()
-
         data = {
-            'host': request.host,
-            'host_name': request.host.split(':')[0],
-            'host_port': request.host_port,
-            'host_url': request.host_url,
-            'path': request.path,
-            'path_url': request.path_url,
+            'host': self.request.host,
+            'host_name': self.request.host.split(':')[0],
+            'host_port': self.request.host_port,
+            'host_url': self.request.host_url,
+            'path': self.request.path,
+            'path_url': self.request.path_url,
         }
 
         return data
@@ -48,19 +135,34 @@ class Bootstrapper(object):
     def config_settings(self):
         """define default settings"""
 
-        request = get_current_request()
-
         data = {
-            'app.software.version': request.registry.settings.get('SOFTWARE_VERSION', ''),
+            'app.software.version': self.registry.settings.get('SOFTWARE_VERSION', ''),
         }
 
         return data
 
+    def resolve_vendor_settings(self):
+        """
+        The selection process will use the first configured vendor as default,
+        after that it will search through the other configured vendors and will
+        select the one which matches the "hostname_matches" pattern
+        on a first come, first serve basis.
+        """
+        selected_vendor_name = self.registry.vendor_settings.vendors[0]
+        selected_vendor_info = self.registry.vendor_settings.vendor[selected_vendor_name]
+        vendor_names = self.registry.vendor_settings.vendors
+        for vendor_name in vendor_names:
+            vendor_info = self.registry.vendor_settings.vendor[vendor_name]
+            if 'hostname_matches' in vendor_info:
+                for hostname_candidate in vendor_info.hostname_matches:
+                    if self.hostname in hostname_candidate:
+                        selected_vendor_info = vendor_info
+                        break
+
+        return selected_vendor_info
+
     def theme_settings(self):
         """define default settings"""
-
-        request = get_current_request()
-        registry = request.registry
 
         data_source_list = [
             '<a href="https://ops.epo.org/" target="_blank" class="incognito pointer">EPO/OPS</a>',
@@ -71,45 +173,28 @@ class Bootstrapper(object):
             '<a href="https://depa.tech/" target="_blank" class="incognito pointer">MTC depa.tech</a>',
         ]
 
-        software_version_label = 'Software release: ' + request.registry.settings.get('SOFTWARE_VERSION', '')
+        software_version_label = 'Software release: ' + self.registry.settings.get('SOFTWARE_VERSION', '')
         software_version_link  = '<a href="https://github.com/ip-tools/ip-navigator" target="_blank" ' \
                                  'class="incognito pointer">{label}</a>'.format(label=software_version_label)
 
-        vendor = registry.application_settings.vendor
+        vendor = self.resolve_vendor_settings()
         data = {
-            'ui.vendor.name': vendor.organization.decode('utf-8'),
-            'ui.vendor.copyright': vendor.copyright_html.decode('utf-8'),
-            'ui.productname': vendor.productname.decode('utf-8'),
-            'ui.productname.rich': vendor.productname_html.decode('utf-8'),
-            'ui.productname.html': vendor.productname_html.decode('utf-8'),
-            'ui.email.purchase': vendor.email_purchase.decode('utf-8'),
-            'ui.email.support': vendor.email_support.decode('utf-8'),
+            'ui.vendor.name': vendor.organization,
+            'ui.vendor.copyright': vendor.copyright_html,
+            'ui.productname': vendor.productname,
+            'ui.productname.rich': vendor.productname_html,
+            'ui.productname.html': vendor.productname_html,
+            'ui.email.purchase': vendor.email_purchase,
+            'ui.email.support': vendor.email_support,
 
             'ui.version': software_version_link,
-            'ui.page.title': 'Patent search', # + ' &nbsp; ' + self.beta_badge,
+            'ui.page.title': vendor.get('page_title', ''), # + ' &nbsp; ' + self.beta_badge,
             'ui.page.subtitle': '',
             'ui.page.footer': 'Data sources: ' + u', '.join(data_source_list),
-
-            'ui.css': {
-                # Page header background
-                '.header-container': {
-                    'background-image': 'url(/static/img/header-small.jpg)',
-                },
-                '.header-inner-left': {
-                    'margin-top': '8px',
-                },
-                '.header-inner-right': {
-                    'margin-top': '10px',
-                },
-
-                '.header-logo': {
-                    'font-size': '24.5px',
-                    'line-height': '40px',
-                    'font-weight': 'bold',
-                },
-
-            }
         }
+
+        if 'stylesheet_uri' in vendor and vendor.stylesheet_uri:
+            data['ui.stylesheet'] = vendor.stylesheet_uri
 
         if self.hostname == 'patentview.ip-tools.io':
             dict_merge(data, {
@@ -117,180 +202,12 @@ class Bootstrapper(object):
                 'ui.productname.rich': '',
             })
 
-        if 'patoffice-navigator' in self.hostname or \
-            self.hostname.endswith('.europatent.net') or \
-            self.hostname.endswith('.patoffice.de'):
-            vendor_color = 'rgba(96, 125, 139, 0.4)'
-            dict_merge(data, {
-                'ui.productname': 'PATOffice Navigator',
-                'ui.productname.rich': '<span class="header-logo">PATOffice Navigator</span> <img width="130" src="/static/vendor/europatent/europatent_logo.png"/>',
-                'ui.header.background_image': None,
-                'ui.page.title': '',
-                'ui.css': {
-
-                    # Header styles
-                    '.header-container': {
-
-                        # Page header background
-                        'background-image': None,
-
-                        #'background-color': vendor_color,
-                        'padding-bottom': '10px',
-                        'border-bottom': '1px solid #aaaaaa',
-
-                        'margin-bottom': '0px',
-
-                        # http://www.cssmatic.com/box-shadow
-                        #'-webkit-box-shadow': '0px 10px 7px 3px {color}'.format(color=vendor_color),
-                        #'-moz-box-shadow': '0px 10px 7px 3px {color}'.format(color=vendor_color),
-                        #'box-shadow': '0px 10px 7px 3px {color}'.format(color=vendor_color),
-                        },
-
-                    '.header-inner-left': {
-                        'margin-top': '35px',
-                    },
-
-                    '.header-logo': {
-                        'font-size': '3em',
-                        'font-weight': 500,
-                        #'font-variant': 'small-caps',
-                        'letter-spacing': '0.05em',
-                        'color': '#333',
-                        'margin-right': '0.6em',
-                        },
-
-                    }
-            })
-
-        if 'inav-web' in self.hostname:
-            vendor_color = '#7ec622'
-            dict_merge(data, {
-                'ui.productname': ' FulltextPRONavigator',
-                'ui.productname.rich': '<div class="header-logo">Navigator</div> <img width="400" src="/static/vendor/ftpro/ftpro_logo.png"/>',
-                'ui.header.background_image': None,
-                'ui.page.title': '',
-                'ui.css': {
-
-                    # Header styles
-                    '.header-container': {
-
-                        # Page header background
-                        'background-image': None,
-                        'background-color': vendor_color,
-                        'margin-bottom': '20px',
-
-                        # http://www.cssmatic.com/box-shadow
-                        '-webkit-box-shadow': '0px 10px 7px 3px {color}'.format(color=vendor_color),
-                        '-moz-box-shadow': '0px 10px 7px 3px {color}'.format(color=vendor_color),
-                        'box-shadow': '0px 10px 7px 3px {color}'.format(color=vendor_color),
-                        },
-
-                    '.header-inner-left': {
-                        'margin-top': '25px',
-                        },
-
-                    '.header-logo': {
-                        #'font-family': 'Arial Black',
-                        'font-size': '3em',
-                        'font-weight': 500,
-                        #'font-variant': 'small-caps',
-                        'letter-spacing': '0.05em',
-                        'color': '#fff',
-                        'margin-right': '0.6em',
-                        'line-height': '50px',
-                        },
-
-                    }
-            })
-
-        if 'patselect' in self.hostname:
-            vendor_color = 'rgba(54,127,179,0.71)'
-            dict_merge(data, {
-                'ui.productname': 'SERVIVA Patselect Navigator',
-                'ui.productname.rich': '<span class="header-logo">Patselect Navigator</span> <img src="/static/vendor/serviva/serviva_logo.png"/>',
-                'ui.header.background_image': None,
-                'ui.page.title': '',
-                'ui.css': {
-
-                    # Header styles
-                    '.header-container': {
-
-                        # Page header background
-                        'background-image': None,
-                        'background-color': vendor_color,
-                        'margin-bottom': '20px',
-
-                        # http://www.cssmatic.com/box-shadow
-                        '-webkit-box-shadow': '0px 10px 7px 3px {color}'.format(color=vendor_color),
-                        '-moz-box-shadow': '0px 10px 7px 3px {color}'.format(color=vendor_color),
-                        'box-shadow': '0px 10px 7px 3px {color}'.format(color=vendor_color),
-                    },
-
-                    '.header-inner-left': {
-                        'margin-top': '25px',
-                    },
-
-                    '.header-logo': {
-                        'font-size': '3em',
-                        'font-weight': 500,
-                        'font-variant': 'small-caps',
-                        'letter-spacing': '0.05em',
-                        'color': '#333',
-                        'margin-right': '0.8em',
-                    },
-
-                }
-            })
-
-        if 'depatech' in self.hostname or 'depa.tech' in self.hostname:
-            vendor_color = '#fff'
-            dict_merge(data, {
-                'ui.productname': 'depa.tech navigator',
-                'ui.productname.rich': '<img src="/static/vendor/mtc/depatech_logo.png"/>',
-                'ui.header.background_image': None,
-                'ui.page.title': '',
-                'ui.css': {
-
-                    # Header styles
-                    '.header-container': {
-
-                        # Page header background
-                        'background-image': None,
-                        'background-color': vendor_color,
-                        'margin-bottom': '20px',
-
-                        # http://www.cssmatic.com/box-shadow
-                        '-webkit-box-shadow': '0px 10px 7px 3px {color}'.format(color=vendor_color),
-                        '-moz-box-shadow': '0px 10px 7px 3px {color}'.format(color=vendor_color),
-                        'box-shadow': '0px 10px 7px 3px {color}'.format(color=vendor_color),
-                        },
-
-                    '.header-inner-left': {
-                        'margin-top': '15px',
-                        },
-                    '.header-inner-right': {
-                        'margin-top': '15px',
-                        },
-
-                    '.header-logo': {
-                        'font-family': 'Ubuntu,Helvetica,Arial,Verdana,sans-serif',
-                        'font-size': '3em',
-                        'font-style': 'normal',
-                        'font-variant': 'normal',
-                        'font-weight': '500',
-                        'font-stretch': 'condensed',
-                        'letter-spacing': '3px',
-                        'color': '#6c0439',
-                        'margin-right': '0.6em',
-                        },
-
-                    }
-            })
-
         return data
 
     def datasource_settings(self):
-        """Return datasource settings accounting for sensible settings like API URI and credentials"""
+        """
+        Return datasource settings while accounting for sensible settings like API URI and credentials.
+        """
         request = get_current_request()
         datasource_settings = SmartBunch.bunchify(request.registry.datasource_settings)
         if 'protected_fields' in datasource_settings:
@@ -391,30 +308,18 @@ class Bootstrapper(object):
         user_modules = params.get('user.modules', [])
         self.hostname = params.get('request.host_name')
 
-        ftpro_allowed_hosts = [
-            'localhost',
-            'offgrid',
-        ]
         params['ftpro_enabled'] =\
-            self.hostname in ftpro_allowed_hosts or \
+            self.hostname in self.development_hosts or \
             'ftpro' in user_modules
 
         # 2.b compute whether datasource "IFI Claims" is enabled
-        ificlaims_allowed_hosts = [
-            'localhost',
-            'offgrid',
-            ]
         params['ifi_enabled'] =\
-            self.hostname in ificlaims_allowed_hosts or \
+            self.hostname in self.development_hosts or \
             'ifi' in user_modules
 
         # 2.c compute whether datasource "MTC depa.tech" is enabled
-        depatech_allowed_hosts = [
-            'localhost',
-            'offgrid',
-            ]
         params['depatech_enabled'] =\
-            self.hostname in depatech_allowed_hosts or\
+            self.hostname in self.development_hosts or\
             'depatech' in user_modules
 
         # 2.d compute whether Google datasource is allowed
@@ -450,9 +355,7 @@ class Bootstrapper(object):
             elif 'localhost' in self.hostname:
                 badge_text = 'localhost'
                 label_kind = 'info'
-            elif 'localhost' in self.hostname:
-                badge_text = 'localhost'
-                label_kind = 'info'
+
             if badge_text and label_kind:
                 params['ui.page.title'] += ' &nbsp; ' + '<div class="label label-{label_kind} label-large do-not-print">{badge_text}</div>'.format(**locals())
 
@@ -474,34 +377,24 @@ class Bootstrapper(object):
         return params
 
 
-class BackboneModelParameterFiddler(object):
+class JavascriptParameterFiddler(object):
 
-    def __init__(self, name):
+    def __init__(self, name, runtime_settings):
         self.name = name
+        self.runtime_settings = runtime_settings
 
     def render(self):
-        """transfer parameters to Backbone model"""
+        """
+        Transfer runtime settings to Javascript
+        """
 
-        boot = Bootstrapper()
-
-        # serialize parameters to json
-        # merge hidden request parameters, e.g. "database=" gets stripped away form site.mako to avoid referrer spam
-        # push current configuration state to browser history
-        tplvars = boot.__dict__.copy()
+        # Serialize parameters to json
+        # Merge hidden request parameters, e.g. "database=" gets stripped away from site.mako to avoid referrer spam
+        # Push current configuration state to browser history
+        tplvars = self.runtime_settings.asdict()
         tplvars['config'] = json.dumps(tplvars['config'])
         tplvars['theme']  = json.dumps(tplvars['theme'])
         javascript_config = 'var navigator_configuration = {config};'.format(**tplvars)
         javascript_theme  = 'var navigator_theme = {theme};'.format(**tplvars)
         payload = '\n'.join([javascript_config, javascript_theme])
         return payload
-
-
-fiddler = BackboneModelParameterFiddler('navigatorConfiguration')
-
-
-def development_mode():
-    registry = get_current_registry()
-    try:
-        return asbool(registry.application_settings.ip_navigator.development_mode)
-    except:
-        return False