+++ CHANGES.rst
@@ -11,6 +11,7 @@ development
 - Use jquery.viewport from npmjs.com
 - IFI Claims: Improve description fulltext display re.
   section “advantageous effects” (missing), embedded lists and embedded drawings
+- Improve handling of global- and runtime configuration settings
 
 2017-09-15 0.136.0
 ==================